### 2022-08-09 #4651 @eloisabentivegna
> 👋 @eloisabentivegna – is there any chance you'd be able to take on an additional submission for JOSS? No worries if not!

Can do!
