### 2022-09-15 87406bb
Update gomamposstoptS.f
