### 2022-09-26 #4800 @editorialbot
[REVIEW]: MG-MAMPOSSt, a Fortran code to test gravity at 2 galaxy-cluster scales
**Submitting author:** @Pizzuti92 (<a href="http://orcid.org/0000-0001-5654-7580">Lorenzo Pizzuti</a>)
**Repository:** https://github.com/Pizzuti92/MG-MAMPOSSt
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.2
