### 2022-10-13 247951f
Update MAM.f
