### 2022-12-13 #4800 @eloisabentivegna
@Pizzuti92, could you comment on what the contributions of your co-authors are to this package? From the repository alone, it seems you are MG-MAMPOSSt's sole developer. You can check JOSS' authorship policy at https://joss.readthedocs.io/en/latest/submitting.html#authorship.
