+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,6 @@
 on: [push] 
 
-jobs: 
+jobs:  
   paper: 
     runs-on: ubuntu-latest 
     name: Paper Draft 
