### 2022-12-19 #4800 @Pizzuti92
@eloisabentivegna the paper Pizzuti et al., 2021 presents a preliminary version where however all main physics of the code and presented is shown along with some forecasts;  with respect to that version, the submitted code ha several minor bugs corrected, a simpler installation and usage procedure implemented, new options for the mass profile (generalized NFW model has been added in standard gravity) and two new models included for the velocity anisotropy profile, characterised by two free parameters which better capture the inner and outer behaviour of the orbits of galaxies.
