### 2022-12-19 #4800 @danielskatz
@eloisabentivegna - just to add a couple of things as another track editor:
1. The preprint on line 202 (https://arxiv.org/abs/2201.07194) shouldn't be an issue - JOSS is fine with preprints, and this seems to be an extended preprint for this JOSS submission
2. The journal paper on line 199 (https://doi.org/10.1093/mnras/stab1727) is more uncertain. The paper seems to focus on algorithm and the science results, as the author says above, but the title of it is "MG-MAMPOSST: a code to"...  I see in the paper "Our main goal is to showcase the code for the aforementioned class of modified gravity models, and present an analysis of statistical degeneracy between the free parameters of the model, which will be essential when testing against real data."  And "In Section 3, we present the MG-MAMPOSST code, its features, and the new modules implemented with respect to the original version of Mamon et al. (2013)."  After reading Section 3.1 of this paper, I do think it focuses on the algorithm much more than the software itself (other than Figure 1 perhaps), so I would lean towards saying that this is not an issue.

Of course, I'm happy to defer to @openjournals/pe-eics and you and the reviewers; this is just my opinion as a non-astrophysicist.
