### 2022-12-19 #4800 @kyleniemeyer
@danielskatz I think this is actually in the @openjournals/aass-eics track. However, we have generally held that existing papers about the software would be duplicate—and a paper explicitly describing the software (i.e., with the software in the name) would seem to be that. Non-JOSS journals would require more focus on the algorithms and results than we do, but the other paper is still about the software itself.
