### 2022-12-19 #4800 @danielskatz
sorry about the track issue, and I agree with you in general, but actually reading the other paper makes me think that its title is misleading and it is not actually about the code, but rather the algorithm and how it can be used.  In any case, yes, we can defer to @openjournals/aass-eics
