### 2022-12-19 #4800 @dfm
@eloisabentivegna, @Pizzuti92 — Thanks for the ping! Taking a look through [the MNRAS paper](https://arxiv.org/abs/2011.15089), I would agree with @danielskatz that I believe that this submission does satisfy our policies for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) and [co-publication](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software), and would be in favor of continuing this review. I'd say that the "code" in the title of the previous publication is used a little "loosely".

@eloisabentivegna — I'd be very happy to chat further (here or offline) if that would be useful. Thanks for checking in!
