### 2022-12-19 #4800 @eloisabentivegna
Thanks, @danielskatz @kyleniemeyer and @dfm for your review of the matter. I concur that the other publication has a different spin that the present one, but was not sure where to draw the line exactly.

Thanks for clarifying what the official position is. I am happy to carry on!
