### 2022-12-20 Pizzuti92/MG-MAMPOSSt#1 @eloisabentivegna
Installation instructions.
The instructions seem to refer to a prior distribution method (a zip file). They should be adapted for users who will obtain the code by cloning the git repository. Also, the documentation paragraph provides instructions that only work if the code has already been installed following the instructions under "Install", so I recommend moving it after this section.

(Part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4800)
