### 2022-12-20 Pizzuti92/MG-MAMPOSSt#2 @eloisabentivegna
Missing updated journal reference/DOI.
Dima & Vernizzi 2017 has now been published (https://journals.aps.org/prd/abstract/10.1103/PhysRevD.97.101302). Please update the article record, including its DOI.

(Part of JOSS review https://github.com/openjournals/joss-reviews/issues/4800)
