### 2022-12-20 Pizzuti92/MG-MAMPOSSt#3 @eloisabentivegna
Missing community guidelines.
As part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/4800), some guidelines for how the community can receive support using the package and contribute extensions must be present. Please add a section (on the README, for instance) explaining how to best interact with MG-MAMPOSSt's developer team.
