### 2022-12-21 Pizzuti92/MG-MAMPOSSt#1 @Pizzuti92
I moved the section "Documentation" after the section "Install" and I added a sentence to explain how to download the code by cloning the git repository
