+++ paper.bib
@@ -511,16 +511,20 @@ archivePrefix = {arXiv},
 }
 
 @article{Dima:2017pwp,
-	author         = {Dima, Alexandru and Vernizzi, Filippo},
-	title          = "{Vainshtein Screening in Scalar-Tensor Theories before
-	and after GW170817: Constraints on Theories beyond
-	Horndeski}",
-	year           = {2017},
-	eprint         = {1712.04731},
-	archivePrefix  = {arXiv},
-	primaryClass   = {gr-qc},
-	SLACcitation   = {%%CITATION = ARXIV:1712.04731;%%}
+  title = {Vainshtein screening in scalar-tensor theories before and after GW170817: Constraints on theories beyond Horndeski},
+  author = {Dima, Alexandru and Vernizzi, Filippo},
+  journal = {Phys. Rev. D},
+  volume = {97},
+  issue = {10},
+  pages = {101302},
+  numpages = {7},
+  year = {2018},
+  month = {May},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevD.97.101302},
+  url = {https://link.aps.org/doi/10.1103/PhysRevD.97.101302}
 }
+
 @article{Crisostomi:2017lbg,
 	author         = {Crisostomi, Marco and Koyama, Kazuya},
 	title          = "{Vainshtein mechanism after GW170817}",
