+++ README.md
@@ -500,7 +500,8 @@ Running ```$ sh script/script_runmam.sh -t``` in this case should produce the [m
 ## Contacts, extension and update
 
 For further information, to report issues and providing suggestions, as well as if you are interested in contributing to the development of the package, please contact Lorenzo Pizzuti (pizzuti@fzu.cz) or Andrea Biviano (andrea.biviano@inaf.it) on behalf of the developers team.
-Since the expression of the gravitational potential is involved only in the solution of the Jeans' equation, new mass models or modified gravity/dark energy parametrizations can be introduced by modifying the routines where the potential enters, which are the functions fa(tlog) and sr2int(alr) in MAM.f. Add a new entry in the condition chain, identified by a new integer number **kmp** which should be string for **M(r)**. 
+
+Extending the code to other mass models is straigthforward. Since the expression of the gravitational potential is involved only in the solution of the Jeans' equation, new mass models or modified gravity/dark energy parametrizations can be introduced by modifying the routines where the potential enters, which are the functions fa(tlog) and sr2int(alr) in MAM.f. Add a new entry in the condition chain, identified by a new integer number **kmp** which should be string for **M(r)**. 
 
 
 
