### 2022-12-21 Pizzuti92/MG-MAMPOSSt#3 @Pizzuti92
I added a new sentence in the last section "Contacts, extension and update":

"For further information, to report issues and providing suggestions, as well as if you are interested in contributing to the development of the package, please contact Lorenzo Pizzuti (pizzuti@fzu.cz) or Andrea Biviano (andrea.biviano@inaf.it) on behalf of the developers team."
