+++ README.md
@@ -8,7 +8,7 @@
 - [Install](#Install)
 - [Overview of usage](#Overview-of-usage)
 - [Test and Tutorial](#Test-and-Tutorial)
-- [Extension and Update](#Extension-and-Update)
+- [Contacts, extension and Update](#Contacts-extension-and-Update)
 ## Introduction
 
 MG-MAMPOSSt is a FORTRAN code that extended the MAMPOSSt algorithm of G. Mamon, A. Biviano and G. Boué - 
