### 2022-12-21 Pizzuti92/MG-MAMPOSSt#4 @eloisabentivegna
Clarify/consolidate list of dependencies.
I have ticked the "Installation instructions" box on the JOSS review https://github.com/openjournals/joss-reviews/issues/4800, but the requirements are currently a bit scattered: there is some information on the README (e.g. for cmake), some in the cmake files, and so on. It would help to make this information more prominent, perhaps in a bulleted list within the "Install" section. In the process, please check that all requirements are spelled out (for instance, I cannot find whether there's a minimum or recommended version for the Fortran compiler).
