### 2022-12-21 Pizzuti92/MG-MAMPOSSt#4 @Pizzuti92
I added a small section in the README:

"###Requirements and dependencies
The following requirements should be satisfied in order to correctly configure and install the code:

-FORTRAN95 compiler or higher
-CMake 3.17.1 or higher
-Python3 or higher. This would be needed for the additional plot.py script, which generates plot at the end of the MG-MAMPOSSt run (see below).
-The software needs some basic math functions included in the libm FORTRAN library."
