### 2022-12-21 Pizzuti92/MG-MAMPOSSt#5 @eloisabentivegna
Enhance the reproducibility of the test.
At the moment, the test provided with the code is not guaranteed to yield identical results to the data provided in test/, due to the MCMC component. This is understandable but not idea, as a user won't in general know if the differences are due to the stochastic nature of the process or a problem with the installation. Would it be possible to design a deterministic test so users can be sure they have a functioning copy of the code? Or at least specify what tolerance the current test results should have? Automating the verification process would also be beneficial, e.g. by providing a diff action in script/script_runmam.sh which checks that the results are correct and issues a "Passed"/"Failed" statement (as opposed to requiring the user to compare data files manually).

(Part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4800)
