### 2022-12-21 Pizzuti92/MG-MAMPOSSt#6 @eloisabentivegna
Provide metadata in the output files.
Some metadata (column headers at a very minimum) would be very beneficial in all output. At the moment, it is not straightforward to identify which quantities (e.g., the lensing forecasts) are output where without sifting through the code.

(Part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4800)
