+++ paper.md
@@ -53,11 +53,11 @@ link-citations: true
 ---
 
 # Summary
-\textsc{MG-MAMPOSSt} is a license-free \textsc{Fortran90} code to perform tests of General Relativity (GR) through the analyses of kinematical data of galaxy clusters. The code solves the Jeans equation, relying on the \textsc{MAMPOSSt} method of @Mamon01. It extends the latter method through new parametrisations of the gravitational potential for general families of gravity theories beyond GR aimed to explain the late-time accelerated expansion of the universe (@riess98, @Perlmutter99).Through appropriate input of projected positions and line-of-sight velocities of cluster's member galaxies, \textsc{MG-MAMPOSSt} reconstructs the cluster mass profile and the velocity anisotropy profile in modified gravity, jointly constraining the kinematics (mass and anisotropy profile) and modified gravity parameters. The code is further supplemented with a new capability to produce weak lensing forecasts for joint kinematic+lensing analyses, offering a valuable tool for studying the nature of gravity at cluster's scales.
+\textsc{MG-MAMPOSSt} is a \textsc{Fortran90} code to perform tests of General Relativity (GR) through the analysis of the kinematical data of galaxy clusters. The code solves the Jeans equation, building on the \textsc{MAMPOSSt} method of @Mamon01. It extends the latter method through new parametrisations of the gravitational potential for general families of gravity theories beyond GR aimed to explain the late-time accelerated expansion of the universe (@riess98, @Perlmutter99).Through appropriate input of the projected positions and line-of-sight velocities of a cluster's member galaxies, \textsc{MG-MAMPOSSt} reconstructs the cluster mass profile and the velocity anisotropy profile in modified gravity, jointly constraining the kinematics (mass and anisotropy profile) and modified gravity parameters. The code is further supplemented with a new capability to produce weak lensing forecasts for joint kinematic+lensing analysis, offering a valuable tool for studying the nature of gravity at cluster's scales.
 
 # Statement of need
 
-In the last two decades, the interest amongst the communities of cosmology and astrophysics in testing dark energy theories at large scales has been grown up significantly. Among the vast range of probes of modified gravity and dark energy models, galaxy clusters offer a powerful laboratory at scales where possible departures from General Relativity should become observable (e.g. @Cataneo19review and references therein). Cluster mass profiles (e.g. @Wilcox15, @Sakstein:2016ggl, @Pizzuti:2016ouw) and the cluster abundance (e.g @Lombriser:2010mp, @Cataneo:2016iav) have been extensively used in the literature to place stringent bounds on some popular classes of extensions of GR. In this context, we developed \textsc{MG-MAMPOSSt}, a \textsc{Fortran90} code capable of performing tests of gravity models with the kinematics of member galaxies in clusters. The code is based upon the original \textsc{MAMPOSSt} method, developed by G. Mamon, A. Biviano and G. Boué (@Mamon01, hereafter MAM13). A public version of \textsc{MAMPOSSt} by G. Mamon can be found at [https://gitlab.com/gmamon/MAMPOSSt](https://gitlab.com/gmamon/MAMPOSSt). Whereas the original code relies on the assumption of a standard Newtonian gravitational potential, \textsc{MG-MAMPOSSt} implements general and viable models of gravity beyond GR, with the aim of placing constraints on their theory space and investigating the essential statistical correlations between model parameters. In addition, the code is capable of producing complementary weak-lensing forecasts for joint kinematics+lensing analyses. \textsc{MG-MAMPOSSt} has been first presented and applied to real galaxy cluster data in @Pizzuti:2017diz, then extended and upgraded by @Pizzuti2021.
[…]
