### 2022-12-21 #4800 @eloisabentivegna
Hi @Pizzuti92, I have made some suggestions for the main text of your paper in the PR https://github.com/Pizzuti92/MG-MAMPOSSt/pull/7. I hope the reasons are clear, but I am happy to elaborate. Could you check that you are OK with the changes, and, if so, merge the PR?
