### 2022-12-21 Pizzuti92/MG-MAMPOSSt#5 @Pizzuti92
I think I can directly implement an automated test script as a flag in script_runmam.sh (eg sh script_runmam.sh - - test) that compares the best fit found by the code with the one expected. To account for the stochastic nature of the process, I can set a tolerance on the value of the likelihood (i.e. if the difference in the log likelihoods best fit from the Montecarlo analysis is less then 1, then the test is considered passed, failed otherwise(. I will develop the test script asap
