### 2022-12-22 f328622
Update script_runmam.sh
