### 2022-12-22 Pizzuti92/MG-MAMPOSSt#5 @Pizzuti92
Dear @eloisabentivegna I added a new part of the script "script_runmam.sh" with the option "-ts" which performs the basic smoke test described in the README. The  "PASSED"/"FAIL" is determined by comparing the best fits (the last two lines in MaxLik.dat") with the expected values. The test is considerd passed if the difference in the likelihood is less then 1. The test requires the additional script "test.py"
