### 2022-12-23 #4800 @eloisabentivegna
@Pizzuti92, thanks for the extended test targetting my question in issue https://github.com/Pizzuti92/MG-MAMPOSSt/issues/5.

Have you added the metadata headers I suggested in https://github.com/Pizzuti92/MG-MAMPOSSt/issues/6? If you could take care of this, I'd be happy to support publication of this work, which entails a final set of actions on your part:
