### 2022-12-23 #4800 @Pizzuti92
@eloisabentivegna sure, this is my task for today... as metadata, I will put headers in the Maxlik.dat file and I can produce a log file where the options used are summarised (e.g. if you have a lensing forecast, mass profile used, best fit parameters)
