### 2022-12-23 Pizzuti92/MG-MAMPOSSt#6 @Pizzuti92
header introduced in MaxLik.dat... Now I will proceed in adding them to all the output files
