+++ gomamposstoptS.f
@@ -851,13 +851,22 @@ c
       open(70,file=fsb,status='unknown')
       open(80,file=fsf,status='unknown')
 
-
-c      needed for old version of plot.py
-c      open(fh,file='Output/log_for_plot',status='unknown')
-c      do i=1,20
-c       write(fh,*) pars(i)
-c      enddo
-c      close(fh)
+      write(30,*) 'record of N(R) model used'
+             
+      write(40,*) 'Binned project N(R) profile'
+      write(40,*)  
+      
+      write(50,*) 'Best fit project N(R) profile'
+      write(50,*) 'R[Mpc], N(R)' 
+       
+      WRITE(60,*) 'Table containing parameters and -log(P)'
+      WRITE(60,*) ''  
+        
+      write(70,*) 'Binned velocity dispersion profile'
+      write(70,*) 'R[Mpc], VDP [km/s], err_up[km/s] err_low [km/s]'
+      
+      write(80,*) 'Fitted velocity dispersion profile'
+      write(80,*) 'R[Mpc], VDP [km/s]'      
       
 c     read system properties
 
@@ -949,7 +958,8 @@ c     MAMPOSSt subroutine
       write(*,*) ' '
 
 c     output a binned number density profile N(R)
-
+      write(40,*) 'Output a binned number density profile N(R)'
+      write(40,*) 'rbin [Mpc], density [1/Mpc^3], err [1/Mpc^3]'
       ibwt=0
       nbins=int(dsqrt(dfloat(nga)))
       do j=1,nga
