### 2022-12-23 Pizzuti92/MG-MAMPOSSt#6 @Pizzuti92
Headers have been introduced in all the output files
