### 2022-12-24 5038eab
Update MAM.f
