### 2022-12-24 363d2ae
Update MAM.f
