### 2022-12-24 Pizzuti92/MG-MAMPOSSt#6 @Pizzuti92
Dear @eloisabentivegna I added a new file, which is generated by running the code, called "Run_info.txt". It contains some information about the run for the user, e.g. if a Monte Carlo run is selected or if the lensing forecast is added to the posterior
