### 2023-01-05 0909106
Update MAM.f
