### 2023-01-05 #4800 @eloisabentivegna
Hi @Pizzuti92! How are things proceeding with the metadata? Should I regenerated the output files with the latest version yet?
