### 2023-01-05 #4800 @Pizzuti92
Dear @eloisabentivegna ,
yes please, if you can generate the new output files with the latest version, there should be headers in all files and the "run_info.txt" file which collects some log info about the run.
