### 2023-01-09 #4800 @eloisabentivegna
Thanks, @Pizzuti92. I have created a PR with some suggestions to improve the header files. If you are happy with the changes, feel free to proceed with the checklist above to prepare the submission for acceptance. But also feel free to tweak my PR to fit your needs.
