### 2023-01-09 #4800 @Pizzuti92
Dear @eloisabentivegna, I fulfilled the points in the checklist:
- New Release made.
Release tag: v1.0.2
Previous releases: v1.0.0, v1.0.1
- Release archived on Zenodo
- Metadata checked
- DOI: 10.5281/zenodo.7515195
For some reasons, the system says the archive is public on Zenodo, but I am not still able to see it searching with the DOI. I immagine that it requires some time to be processed.
