### 2023-01-10 #4800 @Pizzuti92
@eloisabentivegna, just to know, may I run the command "@editorialbot accept" by myself or should I wait further information?
