### 2023-01-10 #4800 @eloisabentivegna
> @eloisabentivegna, just to know, may I run the command "@editorialbot accept" by myself or should I wait further information?

Hi! We need to wait for one of the EiCs to formally accept the submission.
