### 2023-01-10 8b4c0e0
expanding journal macros
