+++ paper.bib
@@ -33,7 +33,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Pizzuti22,
        author = {{Pizzuti}, Lorenzo and {Saltas}, Ippocratis D. and {Umetsu}, Keiichi and {Sartoris}, Barbara},
         title = "{Probing vainsthein-screening gravity with galaxy clusters using internal kinematics and strong and weak lensing}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, gravitational lensing: weak, galaxies: clusters: general, dark energy, dark matter, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2022,
         month = may,
@@ -67,7 +67,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Jain16,
        author = {{Jain}, Rajeev Kumar and {Kouvaris}, Chris and {Nielsen}, Niklas Gr{\o}nlund},
         title = "{White Dwarf Critical Tests for Modified Gravity}",
-      journal = {\prl},
+      journal = {Physical Review Letters},
      keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Solar and Stellar Astrophysics, General Relativity and Quantum Cosmology},
          year = 2016,
         month = apr,
@@ -108,7 +108,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Wilson20,
        author = {{Wilson}, Christopher and {Bean}, Rachel},
         title = "{Testing f (R ) gravity with scale dependent cosmic void velocity profiles}",
-      journal = {\prd},
+      journal = {Physical Review D},
      keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2021,
         month = jul,
@@ -144,7 +144,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Laudato22,
        author = {{Laudato}, Enrico and {Salzano}, Vincenzo and {Umetsu}, Keiichi},
         title = "{Multicomponent DHOST analysis in galaxy clusters}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, gravitational lensing: weak, galaxies: clusters: intracluster medium, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2022,
         month = apr,
@@ -179,7 +179,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mamon01,
    author = {{Mamon}, G.~A. and {Biviano}, A. and {Bou{\'e}}, G.},
     title = "{MAMPOSSt: Modelling Anisotropy and Mass Profiles of Observed Spherical Systems - I. Gaussian 3D velocities}",
-  journal = {\mnras},
+  journal = {Monthly Notices of the RAS},
 archivePrefix = "arXiv",
    eprint = {1212.1455},
  primaryClass = "astro-ph.CO",
@@ -197,7 +197,7 @@ archivePrefix = "arXiv",
    author = {{Riess}, A.~G. and {Filippenko}, A.~V. and {Challis}, P. and 
 	{Clocchiatti}, A. and {Diercks}, A. and {et al.}}, 
     title = "{Observational Evidence from Supernovae for an Accelerating Universe and a Cosmological Constant}",
-  journal = {\aj},
+  journal = {Astronomical Journal},
    eprint = {arXiv:astro-ph/9805201},
  keywords = {COSMOLOGY: OBSERVATIONS, STARS: SUPERNOVAE: GENERAL},
      year = 1998,
@@ -214,7 +214,7 @@ archivePrefix = "arXiv",
[…]
