+++ paper.bib
@@ -31,7 +31,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Pizzuti22,
-       author = {{Pizzuti}, Lorenzo and {Saltas}, Ippocratis D. and {Umetsu}, Keiichi and {Sartoris}, Barbara},
+       author = {{Pizzuti}, L. and {Saltas}, Ippocratis D. and {Umetsu}, Keiichi and {Sartoris}, Barbara},
         title = "{Probing vainsthein-screening gravity with galaxy clusters using internal kinematics and strong and weak lensing}",
       journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, gravitational lensing: weak, galaxies: clusters: general, dark energy, dark matter, Astrophysics - Cosmology and Nongalactic Astrophysics},
@@ -303,7 +303,7 @@ archivePrefix = {arXiv},
 }
 
 @article{Pizzuti:2019wte,
-      author         = {{Pizzuti}, Lorenzo and {Saltas}, Ippocratis D. and {Casas},
+      author         = {{Pizzuti}, L. and {Saltas}, Ippocratis D. and {Casas},
                         Santiago and {Amendol}, Luca and {Biviano}, Andrea},
       title          = "{Future constraints on the gravitational slip with the
                         mass profiles of galaxy clusters}",
