### 2023-01-10 #4800 @eloisabentivegna
> In your checklist above, it's looking like the "Reproducibility" box isn't checked - is that something you still want to look into, or are you happy for this to proceed?

Hi @dfm, I am happy to proceed. I meant to tick that last box after the headers commit, but forgot.

Thanks for checking!
