### 2023-01-10 #4800 @dfm
Many thanks to @ttricco for your review, and to @eloisabentivegna for both editing and reviewing! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

(Note that it can sometimes take up to even a couple of hours for the PDF of the paper to show up on the journal page. I'll check back in later, but ping me if it doesn't show up in the next little bit!)
